# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get categories",
                "operationId": "getCategories",
                "parameters": [
                    {
                        "name": "flag",
                        "in": "query",
                        "description": "flag <b>all</b> for get all categories. flag <b>active</b> for get active categories.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "all",
                                "active"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success,return array of category objects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "store a category",
                "operationId": "bb02f0320882db37f99574196577e446",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token example : 'bearer skfjhskfhksfhkshfkshfkshfkshfkshfse636525xv3535353'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "add this parameter if you want to create a subcategory, otherwise it is not required",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": null
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "category name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Whether a category is active or not This parameter is 1 by default",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": 1,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success,return object of category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/categories/{category_id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get a category by id",
                "operationId": "6902bc43ff20ab7e8c1aaf52393f8471",
                "responses": {
                    "200": {
                        "description": "success,return object of category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "update a category",
                "operationId": "bced945083786a603f6771ff13dc938c",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "if you want to change category parent add this parameter, otherwise it is not required",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "if you want to change category name add this parameter, otherwise it is not required",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "if you want to change category active column add this parameter, otherwise it is not required",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": 1,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "delete a category",
                "operationId": "20febd49287a35c8450e7cb36b2fb2e8",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    }
}